+++ .eslintrc.js
@@ -1,1 +1,1 @@
-module.exports = require('zhouhaifei-code-style').react;
+module.exports = require('./packages/zhouhaifei-code-style/lib/react');