+++ packages/zhouhaifei-code-style/lib/prettier.js
@@ -1,11 +1,3 @@
-const os = require('os');
-
-function isWindows() {
-  return os.type()
-    .toLowerCase()
-    .includes('windows');
-}
-
 module.exports = {
   printWidth: 180,
   tabWidth: 2,
@@ -18,7 +10,7 @@ module.exports = {
   bracketSpacing: true,
   jsxBracketSameLine: false,
   proseWrap: 'never',
-  endOfLine: isWindows() ? 'crlf' : 'lf',
+  endOfLine: 'lf',
   arrowParens: 'always',
   overrides: [
     {